+++ tests/playwright/shiny/bugs/2061-css-js-inclusion/not_include_files_case/customjs.js
@@ -1,0 +1,4 @@
+const newParagraph = document.createElement('p');
+newParagraph.textContent = 'Heyo!';
+document.body.appendChild(newParagraph);
+